Qt5: Remove ModelView
Removing this unused code for now.  This might be reverted in the
future if things change.  My suggestion would be to replace it
with an OSG based one and have one single 3D engine used consistently.

I investiaged briefly using Qt3d but that has been deprecated for the
time being for lack of a maintainer.  Given that it doesn't currently
make sense to try to port modelview as this would already be another
set of libraries we would have to install.

http://www.mail-archive.com/interest@qt-project.org/msg04398.html