Filters: process takes the filter driver
Right now we don't require storing a pointer to the filter driver
in the opaque filter instance handle, which is probably good.  That
means, though, that we need to pass around the filter driver generally
with the instance id.

Previously the SE(3)+ infrastructure method took a specific sub-driver
but that was an architectural mistake since that the process method is
a generic API method and should not use a specific structure.