EKF: remove baro_offset tracking
The GPS wander is a lot more than the baro for the
time periods we are interested in so the simple low
pass filter currently applied isn't a good model.

In the long run trying to run some medium time scale
state estimate to track the GPS altitude bias and that
is centered around zero and a longer time scale bias
for the baro might be a better approach.