VTOLPath: begin FSM framework for simple goals
This begins implementing a framework where there are FSMs
defined to achieve certain goals (e.g. fly to location and
land), which provides a layer of indirection between the
PathDesired UAVO and the actual navigation.

Most of the FSM logic to execute RTH is implemented. However,
none of the control algorithms are connected up yet.