EEG: fix the calculation of the impedance
I was computing the Fourier component incorrectly and multiplying
by 2 instead of dividing by 2. Checking in matlab, this new version
produces the right value as well as comparing the amplitude on the
scope versus calculated on board.

The factor of sqrt(2) was needed to get the values more correct and I
think comes from the fact this is a square wave and not sinusoidal.